Added registration context provider
