+++ login-workflow/example/src/navigation/index.tsx
@@ -9,7 +9,7 @@ import { WorkflowCardWrapper } from '../screens/WorkflowCardWrapper';
 import RegistrationProviderExample from '../screens/RegistrationProviderExample';
 import I18nExample from '../screens/I18nExample';
 import RegistrationContextExample from '../screens/RegistrationContextExample';
-import { AccountDetailsScreen } from '../screens/DemoAccountDetails';
+import DemoAccountDetails from '../screens/DemoAccountDetails';
 
 const Drawer = createDrawerNavigator();
 
@@ -47,7 +47,7 @@ export const MainRouter = (): any => (
             <RootStack.Screen name="I18nExample" component={I18nExample} />
             <RootStack.Screen name="RegistratonContextExample" component={RegistrationContextExample} />
             <RootStack.Screen name="RegistrationProviderExample" component={RegistrationProviderExample} />
-            <RootStack.Screen name="AccountDetailsBaseExample" component={AccountDetailsScreen} />
+            <RootStack.Screen name="AccountDetailsBaseExample" component={DemoAccountDetails} />
         </Drawer.Navigator>
     </NavigationContainer>
 );
